RSB-9 refactor: modify component structure and path
